Revert "Merge branch 'Develop' of https://github.com/implementacaoteste/TEC.2022.1.108.Gestao-de-hoteis into Develop"
This reverts commit c7b3ee9cc35978a9cbb854134e8d84607bb73b4b, reversing
changes made to 47f30a1488992c058a60d07c0524e7fc4e48f300.

+++ WindowsFormsPrincipal/Models/Relatorio.cs
@@ -19,8 +19,8 @@ namespace Models
         public string Pagamento { get; set; }
         public string Numero_Quarto { get; set; }
         public string Descricao_Quarto { get; set; }
-        public decimal Valor_Total { get; set; }
-        public decimal Valor_Diaria { get; set; }
+        public double Valor_Total { get; set; }
+        public double Valor_Diaria { get; set; }
         public DateTime Data_Checkin { get; set; }
         public DateTime Data_Checkout { get; set; }
     }
